% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/los.R
\name{los}
\alias{los}
\title{Hauptfunktion zum Starten des Lernspiels}
\usage{
los(hausis = 7)
}
\arguments{
\item{hausis}{An welchen Wochentagen soll gefragt werden, ob die
Hausaufgaben schon erledigt wurden?}
}
\description{
Wie der Name schon sagt, startet diese Funktion das Lernspiel. Es ist rein
textbasiert, d.h. es läuft in der Konsole ab und erwartet immer wieder
Tastatureingaben, die mit der Eingabetaste abzuschließen sind.
}
