% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiel_Datum.R
\name{spiel_Datum}
\alias{spiel_Datum}
\title{Fragen zu Datum und Uhrzeit}
\usage{
spiel_Datum(name)
}
\arguments{
\item{name}{Der Name des Spielers.}
}
\description{
Welcher Wochentag ist heute und dann übermorgen? Welchen Monat haben wir und
welchen Tag im Monat? Wieviel Uhr ist es?
}
