% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiel_Detektiv.R
\name{spiel_Detektiv}
\alias{spiel_Detektiv}
\title{Detektivspiel mit verschl체sselten S채tzen}
\usage{
spiel_Detektiv(name)
}
\arguments{
\item{name}{Der Name des Spielers.}
}
\description{
In diesem Spiel sollen verschl체sselte S채tze entziffert werden. Die
Buchstaben eines Satzes sind jeweils durch Zahlen ersetzt, entsprechend des
Index im Alphabet.
}
