% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiel_Rechnen.R
\name{spiel_Rechnen}
\alias{spiel_Rechnen}
\title{Rechenaufgaben mit Plus und Minus}
\usage{
spiel_Rechnen(name, minimum = 0, maximum = 20, zahlengrenze = 9)
}
\arguments{
\item{name}{Der Name des Spielers.}

\item{minimum, maximum}{Erlaubter Ergebnisbereich.}

\item{zahlengrenze}{Größte Zahl mit der gerechnet werden soll.}
}
\description{
In diesem Spiel wird eine selbst zu wählende Anzahl von Rechenaufgaben
gestellt. Diese enthalten jeweils einzelne oder verkettete Additionen oder
Subtraktionen der natürlichen Zahlen von 0 bis 9. Der mögliche
Ergebnisbereich wird vorher festgelegt (standardmäßig von 0 bis 20).
}
