% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiel_Tiere.R
\name{spiel_Tiere}
\alias{spiel_Tiere}
\title{Fehlende Buchstaben ergänzen}
\usage{
spiel_Tiere(name)
}
\arguments{
\item{name}{Der Name des Spielers.}
}
\description{
In diesem Spiel sollen fehlende Buchstaben in Wörtern ergänzt werden.
Derzeit sind alle Wörter Tierarten.
}
